scraper/requirements.txt
requests
beautifulsoup4
pytz
pyairtable
python-dotenv